Adicionar dependentes, team members

+++ app/Http/Controllers/EmployeeController.php
@@ -2,10 +2,12 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Support\Facades\Auth;
 use App\Models\Tenant\CompanyBranch;
 use App\Models\Tenant\CompanyBranchEmployee;
 use App\Models\Tenant\CompanyGroup;
 use App\Models\Tenant\Employee;
+use App\Models\Tenant\Position;
 use App\Models\User;
 use Illuminate\Http\Request;
 
@@ -29,9 +31,24 @@ class EmployeeController extends Controller
         $branches = CompanyBranch::all();
         $users = User::role('employee')->get();
         $userIds = $users->pluck('id')->toArray();
+        $positions = Position::all(); // Carrega todas as posições
         // Consultar os funcionários com base nos user_ids
         $employees = Employee::whereIn('user_id', $userIds)->get();
-        return view('employee.create', compact('groups', 'branches', 'employees'));
+        return view('employee.create', compact('groups', 'branches', 'employees', 'positions'));
+    }
+
+    public function dependents($employeeId)
+    {
+        $employee = Employee::findOrFail($employeeId);
+        $dependents = $employee->dependents; // Assumindo que a relação 'dependents' está definida no modelo Employee
+        return view('employee.dependents', compact('employee', 'dependents'));
+    }
+
+    public function team($employeeId)
+    {
+        $employee = Employee::findOrFail($employeeId);
+        $teamMembers = $employee->teamMembers;
+        return view('employee.team', compact('employee', 'teamMembers'));
     }
 
     /**
@@ -39,11 +56,13 @@ class EmployeeController extends Controller
      */
     public function store(Request $request)
     {
-        $user = User::create($request->all());
+        $data_user = $request->all();
+        $data_user['company_id'] = Auth::user()->company_id;
+        $user = User::create($data_user);
 
         $data = $request->all();
         $data['user_id'] = $user->id;
-        $data['manager'] = (isset($data['manager']) && $data['manager'] === 'ON');
+        $data['manager'] = (isset($data['manager']) && $data['manager'] === 'on');
         if($data['manager'])
         {
             $user->assignRole('manager');
@@ -89,9 +108,10 @@ class EmployeeController extends Controller
         $branches = CompanyBranch::all();
         $users = User::role('employee')->get();
         $userIds = $users->pluck('id')->toArray();
+        $positions = Position::all(); // Carrega todas as posições
         // Consultar os funcionários com base nos user_ids
         $employees = Employee::whereIn('user_id', $userIds)->get();
-        return view('employee.edit', compact('employee','groups', 'branches', 'employees'));
+        return view('employee.edit', compact('employee','groups', 'branches', 'employees', 'positions'));
     }
 
     /**
@@ -100,9 +120,14 @@ class EmployeeController extends Controller
     public function update(Request $request, string $id)
     {
         $data = $request->all();
-        $data['manager'] = (isset($data['manager']) && $data['manager'] === 'ON');
-
+        $data['manager'] = (isset($data['manager']) && $data['manager'] === 'on');
+        
         $employee = Employee::find($id);
+        if($data['manager'])
+        {
+            $user = User::where('email', $employee->email)->first();
+            $user->assignRole('manager');
+        }
         $employee->update($data);
 
         return redirect()->route('employee.index');
@@ -114,7 +139,21 @@ class EmployeeController extends Controller
     public function destroy(string $id)
     {
         $employee = Employee::find($id);
+        
+        if (!$employee) {
+            return redirect()->route('employee.index')->with('error', 'Employee not found.');
+        }
+
+        // Busca o usuário associado ao employee
+        $user = User::where('email', $employee->email)->first();
+
+        // Verifica se o usuário existe e o exclui
+        if ($user) {
+            $user->delete();
+        }
+
         $employee->delete();
+
         return redirect()->route('employee.index');
     }
 }
